+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - プロジェクトの返金時設定を削除
 - オファーのアドオンに対しても利用可能アプリケーション設定を適用
 - 注文ステータス変更の冪等性担保
+- 注文取引のカスタマー情報をobject.customerにセットするように調整
 
 ### Deprecated